[Feature] 패키지, artifactid 변경 코드 디렉토리 위치 수정
