--- wise-office-server/backoffice/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'backoffice'
