--- wise-office-server/backoffice/src/main/java/wise/backoffice/backoffice/BackofficeApplication.java
@@ -1,13 +1,0 @@
-package wise.backoffice.backoffice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackofficeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackofficeApplication.class, args);
-	}
-
-}
