--- wise-office-server/backoffice/src/main/java/wise/backoffice/backoffice/TestController.java
@@ -1,16 +1,0 @@
-package wise.backoffice.backoffice;
-
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.GetMapping;
-
-
-@RestController
-public class TestController {
-    
-
-    @GetMapping("/health")
-    public String getMethodName() {
-        return "status OK";
-    }
-    
-}
