--- wise-office-server/backoffice/src/main/java/wise/backoffice/backoffice/domain/member/entity/MemberRoleType.java
@@ -1,5 +1,0 @@
-package wise.backoffice.backoffice.domain.member.entity;
-
-public enum MemberRoleType {
-    MASTER, MANAGER, WORKER;
-}
