--- wise-office-server/backoffice/src/test/java/wise/backoffice/backoffice/BackofficeApplicationTests.java
@@ -1,13 +1,0 @@
-package wise.backoffice.backoffice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackofficeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
