+++ wise-office-server/src/main/java/kr/co/wise/BackofficeApplication.java
@@ -1,0 +1,13 @@
+package kr.co.wise;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackofficeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackofficeApplication.class, args);
+	}
+
+}
