+++ wise-office-server/src/main/java/kr/co/wise/TestController.java
@@ -1,0 +1,16 @@
+package kr.co.wise;
+
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.GetMapping;
+
+
+@RestController
+public class TestController {
+    
+
+    @GetMapping("/health")
+    public String getMethodName() {
+        return "status OK";
+    }
+    
+}
