+++ wise-office-server/src/main/java/kr/co/wise/domain/member/entity/MemberRoleType.java
@@ -1,0 +1,5 @@
+package kr.co.wise.domain.member.entity;
+
+public enum MemberRoleType {
+    MASTER, MANAGER, WORKER;
+}
