+++ wise-office-server/src/test/java/kr/co/wise/BackofficeApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.wise;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackofficeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
